feat: add train_size parameter to run_gpu_umap for subset training
introduce an option to perform the UMAP on a limited set of points. but DBSCAN and plotting shall use the full dataset.

the trained umap reducer shall be used to map the remaining points into the lower dimensional space.

doesn't help with the breakdown of umap at 6_300 points per file